+++ server/src/persistence/schema.rs
@@ -30,6 +30,14 @@ table! {
 }
 
 table! {
+    loadout (id) {
+        id -> Integer,
+        character_id -> Integer,
+        items -> Text,
+    }
+}
+
+table! {
     stats (character_id) {
         character_id -> Integer,
         level -> Integer,
@@ -42,6 +50,7 @@ table! {
 
 joinable!(body -> character (character_id));
 joinable!(inventory -> character (character_id));
+joinable!(loadout -> character (character_id));
 joinable!(stats -> character (character_id));
 
-allow_tables_to_appear_in_same_query!(body, character, inventory, stats);
+allow_tables_to_appear_in_same_query!(body, character, inventory, loadout, stats,);